code/requirements.txt
transformers
peft
accelerate
safetensors
bitsandbytes
datasets
ms-swift